refactor(frontend): split profile settings components
